+++ modules/transport-grpc/src/main/java/org/opensearch/transport/grpc/proto/request/search/sort/SortBuilderProtoUtils.java
@@ -8,7 +8,7 @@
 package org.opensearch.transport.grpc.proto.request.search.sort;
 
 import org.opensearch.core.xcontent.XContentParser;
-import org.opensearch.protobufs.SortOptions;
+import org.opensearch.protobufs.SortCombinations;
 import org.opensearch.search.sort.FieldSortBuilder;
 import org.opensearch.search.sort.ScoreSortBuilder;
 import org.opensearch.search.sort.SortBuilder;
@@ -39,31 +39,9 @@ public class SortBuilderProtoUtils {
      * @throws IllegalArgumentException if invalid sort combinations are provided
      * @throws UnsupportedOperationException if sort options are not yet supported
      */
-    public static List<SortBuilder<?>> fromProto(List<SortOptions> sortProto) {
+    public static List<SortBuilder<?>> fromProto(List<SortCombinations> sortProto) {
         List<SortBuilder<?>> sortFields = new ArrayList<>(2);
-
-        for (SortOptions sortOptions : sortProto) {
-            switch (sortOptions.getSortOptionsCase()) {
-                case STRING:
-                    String name = sortOptions.getString();
-                    sortFields.add(fieldOrScoreSort(name));
-                    break;
-                case SORT_OPTIONS_SCORE:
-                    sortFields.add(new ScoreSortBuilder());
-                    break;
-                case SORT_OPTIONS_DOC:
-                    sortFields.add(new FieldSortBuilder("_doc"));
-                    break;
-                case FIELD_SORT:
-                    // Handle field sort - this is more complex and would need FieldSort parsing
-                    throw new UnsupportedOperationException("Field sort not implemented yet");
-                case SORT_OPTIONS_ONE_OF:
-                    throw new UnsupportedOperationException("Sort options oneof not supported yet");
-                default:
-                    throw new IllegalArgumentException("Invalid sort options provided: " + sortOptions.getSortOptionsCase());
-            }
-        }
-        return sortFields;
+        throw new UnsupportedOperationException("sort not supported yet");
     }
 
     /**